Conflicts:
	Build/FwBuildTasks.dll
	Build/Src/FwBuildTasks/TestTask.cs
	Build/Src/FwBuildTasks/Unitpp.cs
	Build/TestTimeoutValues.xml
	Build/Windows.targets
	Build/mkall.targets
	Lib/src/PhraseTranslationHelper/PhraseTranslationHelperTests/Properties/AssemblyInfo.cs
	Lib/src/PhraseTranslationHelper/Properties/AssemblyInfo.cs
	Lib/src/UIAdapters/DotNetBarAdapter/AssemblyInfo.cs
	Lib/src/UIAdapters/DotNetBarAdapter/DNBAdapter.cs
	Lib/src/UIAdapters/SilSidePaneAdapter/AssemblyInfo.cs
	Lib/src/UIAdapters/SilSidePaneAdapter/CustomDropDown.cs
	Lib/src/UIAdapters/SilSidePaneAdapter/ToolStripAdapter.cs
	Lib/src/UIAdapters/SilSidePaneAdapter/UIAdaptersTestClient/AssemblyInfo.cs
	Lib/src/UIAdapters/ToolStripAdapter/AssemblyInfo.cs
	Lib/src/UIAdapters/ToolStripAdapter/CustomDropDown.cs
	Lib/src/UIAdapters/ToolStripAdapter/SideBarItemAreaRenderer.cs
	Lib/src/UIAdapters/ToolStripAdapter/ToolStripAdapter.cs
	Lib/src/UIAdapters/ToolStripAdapter/UIAdaptersTestClient/AssemblyInfo.cs
	Lib/src/UIAdapters/ToolStripAdapter/UIAdaptersTestClient/Main.cs
	Lib/src/UIAdapters/ToolStripAdapter/UIAdaptersTestClient/MainWindow.cs
	Lib/src/XCoreAdapter-SilSidePane/AssemblyInfo.cs
	Lib/src/XCoreAdapter-SilSidePane/PaneBar.cs
	Lib/src/XCoreAdapter-SilSidePane/PanelButton.cs
	Lib/src/XCoreAdapter-SilSidePane/PanelMenu.cs
	Lib/src/XCoreAdapter/AssemblyInfo.cs
	Lib/src/XCoreAdapter/PaneBar.cs
	Lib/src/XCoreAdapter/PanelButton.cs
	Lib/src/XCoreAdapter/PanelMenu.cs
	Lib/src/XCoreAdapterTestApp/XCoreAdapterTestApp/Properties/AssemblyInfo.cs
	Src/Common/Controls/DetailControls/DetailControlsTests/DataTreeTests.cs
	Src/Common/Controls/XMLViews/XMLViewsTests/TestXmlBrowseView.cs
	Src/Common/CoreImpl/DirectoryFinder.cs
	Src/Common/FieldWorks/FieldWorks.cs
	Src/Common/FieldWorks/FieldWorks.csproj
	Src/Common/Filters/FiltersTests/WordformFiltersTests.cs
	Src/Common/FwPrintLayoutComponents/HFSetView.cs
	Src/Common/FwRemoteDatabaseConnectorService/FwRemoteDatabaseConnectorService.cs
	Src/Common/FwRemoteDatabaseConnectorService/Program.cs
	Src/Common/FwRemoteDatabaseConnectorService/ProjectInstaller.cs
	Src/Common/FwRemoteDatabaseConnectorService/Properties/AssemblyInfo.cs
	Src/Common/FwUtils/FwUtilsTests/UpgradeUserSettingsTests.cs
	Src/Common/PrintLayout/PrintLayoutTests/PrintLayoutTestBase.cs
	Src/Common/SimpleRootSite/SimpleRootSiteTests/IbusRootSiteEventHandlerTests.cs
	Src/FDO/DomainServices/ClientServerServices.cs
	Src/FDO/DomainServices/WritingSystemServices.cs
	Src/FDO/FDOTests/ClientServerServicesTests.cs
	Src/FDO/FDOTests/ClientServerTests.cs
	Src/FDO/FDOTests/Db4oServerInfoTests.cs
	Src/FDO/FDOTests/PersistingBackendProviderTests.cs
	Src/FDO/FDOTests/PersistingLayerTests.BEPPortTests.cs
	Src/FDO/Infrastructure/Impl/ClientIPServerSink.cs
	Src/FDO/Infrastructure/Impl/ClientIPServerSinkProvider.cs
	Src/FDOBrowser/CustomFields.cs
	Src/FDOBrowser/Db4oToXml.cs
	Src/FDOBrowser/FileInOutChooser.cs
	Src/FdoUi/PartOfSpeechUi.cs
	Src/FwCoreDlgs/BackupRestore/ChangeDefaultBackupDir.cs
	Src/FwCoreDlgs/CnvtrTypeComboItem.cs
	Src/FwCoreDlgs/DeleteWritingSystemWarningDialog.cs
	Src/FwCoreDlgs/FwCoreDlgControls/ConfigSenseLayout.cs
	Src/FwCoreDlgs/FwFindReplaceDlg.cs
	Src/FwCoreDlgs/MoveOrCopyFilesDlg.cs
	Src/FwCoreDlgs/PicturePropertiesDialog.cs
	Src/FwCoreDlgs/WritingSystemPropertiesDialog.cs
	Src/FwParatextLexiconPlugin/ChooseFdoProjectForm.cs
	Src/FwParatextLexiconPlugin/ParatextLexiconPluginRegistryHelper.cs
	Src/LexText/Interlinear/ConcordanceWordList.cs
	Src/LexText/Interlinear/ITextDllTests/ComboHandlerTests.cs
	Src/LexText/Interlinear/ITextDllTests/SandboxBaseTests.cs
	Src/LexText/Interlinear/InterAreaBookmark.cs
	Src/LexText/Interlinear/InterlinDocForAnalysis.cs
	Src/LexText/Interlinear/ParseIsCurrentFixer.cs
	Src/LexText/Interlinear/Sandbox.cs
	Src/LexText/LexTextControls/IFwExtension.cs
	Src/LexText/LexTextControls/InsertRecordDlg.cs
	Src/LexText/LexTextControls/RecordDlgListener.cs
	Src/LexText/LexTextDll/Db4oSendReceiveDialog.Designer.cs
	Src/LexText/LexTextDll/Db4oSendReceiveDialog.cs
	Src/LexText/LexTextDll/FlexStylesXmlAccessor.cs
	Src/LexText/Lexicon/LexEntryMenuHandler.cs
	Src/LexText/Lexicon/RecordReferenceVectorLauncher.cs
	Src/LexText/ParserUI/TryAWordDlg.cs
	Src/MasterVersionInfo.txt
	Src/TE/OxesIO/OxesIOTests/Properties/AssemblyInfo.cs
	Src/TE/OxesIO/Properties/AssemblyInfo.cs
	Src/TE/TeDialogs/TeDialogsTests/ToolsOptionsDialogTests.cs
	Src/TE/TeEditing/PromptSelectionRestorer.cs
	Src/TE/TeEditorialChecks/CheckGridListSorter.cs
	Src/TE/TeEditorialChecks/CheckingErrorCommentDlg.cs
	Src/TE/TeEditorialChecks/TeEditorialChecksTests/CheckGridListSorterTests.cs
	Src/TE/TeImportExport/ExportDefinitions.cs
	Src/TE/TeImportExport/ImportXmlDialog.cs
	Src/TE/TeImportExport/TeImportExportTests/ExportTests/XmlNoteCategoryTests.cs
	Src/TeDll/ScrTextListSelectionForm.cs
	Src/TeDll/ScrTextListSelectionForm.designer.cs
	Src/TeDll/TrainingAvailable.cs
	Src/Utilities/BasicUtils/SimpleLogger.cs
	Src/XCore/HtmlViewer.cs
	Src/XCore/SilSidePaneAdapter/CustomDropDown.cs
	Src/XCore/SilSidePaneAdapter/ToolStripAdapter.cs
	Src/XCore/xCoreInterfaces/RecordFilterListProvider.cs
	Src/XCore/xCoreOpenSourceAdapter/PaneBar.cs
	Src/xWorks/DTMenuHandler.cs
	Src/xWorks/DictionaryDetailsView/ListOptionsView.resx
	Src/xWorks/DictionaryPublicationDecorator.cs
	Src/xWorks/FwXWindow.cs
	Src/xWorks/xWorksStrings.resx
	Src/xWorks/xWorksTests/xWorksTests.csproj

+++ Build/Src/FwBuildTasks/TestTask.cs
@@ -104,7 +104,20 @@ namespace FwBuildTasks
 				{
 					try
 					{
-						process.Kill();
+						bool testLooksToHaveFinished;
+						lock(LockObject)
+						{
+							testLooksToHaveFinished = !string.IsNullOrEmpty(m_TestLog.FindLast(line => line.Contains(GetTestsCompletedString())));
+						}
+						if(testLooksToHaveFinished)
+						{
+							Log.LogMessage(MessageImportance.Normal, "Tests for {0} timed out, but appear to have finished. Giving 2 seconds for log to be written.", TestProgramName);
+							Thread.Sleep(2000);
+							// If the tests completed, don't fail with a timeout; if missing Dispose calls caused NUnit to hang, that should be caught as a different failure.
+							fTimedOut = false;
+						}
+						if(!process.HasExited) // If the tests looked to have finished the process might have exited normally by now
+							process.Kill(); // This will set the exit code to -1 and the suite will be added as a failed suite below
 					}
 					catch
 					{
@@ -167,6 +180,8 @@ namespace FwBuildTasks
 			// test or if we get a timeout we want to fail the build.
 			return retVal;
 		}
+
+		protected abstract string GetTestsCompletedString();
 
 		/// <summary>
 		/// Starts the process and handles errors.
